refactor(tests): consolidate OpenAPI response tests as requested by @leandrodamascena
- Moved all tests from test_openapi_response_fields.py to existing test_openapi_responses.py
- Removed duplicate test file to improve code organization
- All 21 tests pass (12 original + 9 consolidated)
- Addresses reviewer feedback for better test file structure
